+++ src/main/java/ch/heigvd/pclist/services/dao/CpuDAOLocal.java
@@ -16,7 +16,7 @@ public interface CpuDAOLocal {
 
     List<Cpu> get(List<Long> idList);
 
-    List<Cpu> get(int pageSize, int pageIndex);
+    List<Cpu> get(long pageSize, long pageIndex);
 
     List<Cpu> get();
